[metrics] M1 axis-metric integrity + M0 lint/DiCE consolidation
- Add Axis B/D adversarial tests (test_axis_b.py, test_axis_d.py); every
  Axis A-D metric now has a green adversarial test.
- Ground Axis A-D metrics in literature (BibTeX in references_verified.md).
- Remove DiCE (DiCECF): tabular method, never wired to Phase 03. Drop the
  module, exports, unit tests, the dice-ml dependency, and docstring mentions.
- Fix ruff config (drop removed W503 selector; move select/ignore under
  [tool.ruff.lint]) and isort config typo (multi_line_mode ->
  multi_line_output) - both linters were failing CI. Apply the resulting
  ruff autofixes across the tree.

+++ causaltemp_xai/benchmarks/generator.py
@@ -58,7 +58,6 @@ import numpy as np
 
 from causaltemp_xai.benchmarks.mechanisms import LinearMechanism, MLPMechanism
 
-
 _NOISE_TYPES = ("laplace", "uniform", "gaussian")
 
 #: Std-dev for the Gaussian innovation branch (M4/H5 negative-control
@@ -605,7 +604,7 @@ class RegimeSwitchNlinearSCMT:
         note for how ``"mechanism"`` (regime 1) should be interpreted.
         """
         total_T = self.T + burn_in
-        switch_t_abs = burn_in + int(round(self.switch_frac * self.T))
+        switch_t_abs = burn_in + round(self.switch_frac * self.T)
         X_full = np.zeros((self.N, total_T, self.k))
 
         self._roll(X_full, np.arange(self.N), total_T, switch_t_abs)
